Refactor EditProductModal into modular components
