+++ src/pages/Sites/SiteSettings/Catalog/Products/components/EditProductModal/index.js
@@ -1,0 +1,1 @@
+export { default } from './EditProductModal'
